Add files via upload
This Jupyter Notebook contains the implementation of the final machine learning model for predicting heart disease, as developed in the CardioVision AI project. The notebook includes all necessary steps from data preprocessing to model training, evaluation, and testing. It leverages the TensorFlow library to build and optimize a neural network that achieves high predictive accuracy. The notebook is structured to be easily executable, providing detailed comments and explanations for each step to ensure reproducibility and clarity. This final model demonstrates superior performance with optimized hyperparameters and robust generalization capabilities.