Add files via upload
This Jupyter Notebook contains the code for the parameter optimization section of the CardioVision AI project. It details the process of refining hyperparameters for the heart disease prediction model, including regularization parameters and learning rates. The notebook employs systematic grid searches and manual adjustments to identify the optimal values that enhance model performance. Utilizing TensorFlow and other essential libraries, this notebook ensures robust model training and evaluation, ultimately leading to a highly accurate and reliable predictive model. Detailed explanations and visualizations are provided to illustrate the optimization process and its impact on model performance.